Version 5.20

+++ pcican2/pcicanII.sh
@@ -72,7 +72,12 @@ fi
 #
 case "$1" in
    start)
-      /sbin/$kv_module_install kvpcicanII || exit 1
+      MODULE_INSTALL_OUT=$(/sbin/$kv_module_install kvpcicanII 2>&1)
+      MODULE_INSTALL_RES=$?
+      if [ $MODULE_INSTALL_RES -ne 0 ] ; then
+        $LOG -t $0 $MODULE_INSTALL_OUT
+        exit 1
+      fi
       nrchan=`cat /proc/pcicanII | grep 'total channels' | cut -d \  -f 3`
       major=`cat /proc/devices | grep 'pcicanII' | cut -d \  -f 1`
       rm -f /dev/pcicanII*
